# workflow/openapi.yaml
...
openapi: 3.1.0
info:
  title: GW/AP Debug Platform Workflow API
  version: 0.9.0
servers:
- url: http://localhost:8000/api/v1
...
paths:
...
  /knowledge-curations:
    get:
      tags:
      - knowledge-curation
      summary: List Curation Sessions
      operationId: listKnowledgeCurations
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Curation Sessions Api V1 Knowledge Curations Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      x-agent-callable: true
      description: List caller-owned case-extraction sessions. ADMIN, EXPERT and ENGINEER may contribute; managers
        gain review access to another source session only after its associated contribution is submitted.
    post:
      tags:
      - knowledge-curation
      summary: Create Curation Session
      operationId: createKnowledgeCuration
      parameters:
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_create_curation_session_api_v1_knowledge_curations_post'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Create Curation Session Api V1 Knowledge Curations Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      x-agent-callable: true
      description: Create an owner-scoped case-extraction session using an explicitly selected visible Chat profile
        or the personal/shared default. Bind the initiating user and configuration fingerprint for worker revalidation.
        New-session egress consent defaults to true; explicit false remains false.
  /knowledge-curations/{session_id}:
    get:
      tags:
      - knowledge-curation
      summary: Get Curation Session
      operationId: getKnowledgeCuration
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Curation Session Api V1 Knowledge Curations  Session Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      x-agent-callable: true
      description: Read an owned extraction session. ADMIN/EXPERT may read a source session associated with a submitted
        contribution for review; another user private model details are not exposed.
  /knowledge-curations/{session_id}/chat:
    post:
      tags:
      - knowledge-curation
      summary: Chat With Curation Session
      operationId: refineKnowledgeCuration
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeCurationChatRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Chat With Curation Session Api V1 Knowledge Curations  Session Id  Chat Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      x-agent-callable: true
      description: The extraction owner may converse using the session-bound Chat profile to create a new immutable
        draft version. Managers correct another user submitted proposal through contribution review-chat instead.
  /knowledge-curations/{session_id}/confirm:
    post:
      tags:
      - knowledge-curation
      summary: Confirm Curation
      operationId: confirmKnowledgeCurationDraft
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeCurationConfirmRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Confirm Curation Api V1 Knowledge Curations  Session Id  Confirm Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      x-agent-callable: true
      description: The extraction owner confirms the exact draft version to create an inactive knowledge DRAFT and
        an owned contribution proposal. This does not submit, approve, or publish the proposal.
...
  /workbench/knowledge:
    get:
      tags:
      - workbench
      summary: Knowledge
      operationId: knowledge_api_v1_workbench_knowledge_get
      parameters:
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      - VIEWER
      x-agent-callable: true
...
components:
  schemas:
...
    Body_create_curation_session_api_v1_knowledge_curations_post:
      properties:
        files:
          items:
            type: string
            contentMediaType: application/octet-stream
          type: array
          title: Files
        relative_paths_json:
          type: string
          title: Relative Paths Json
          default: '[]'
        title_hint:
          type: string
          title: Title Hint
          default: ''
        category_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Category Id
        device_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Type
        device_model:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Model
        firmware_range:
          anyOf:
          - type: string
          - type: 'null'
          title: Firmware Range
        module:
          anyOf:
          - type: string
          - type: 'null'
          title: Module
        trust_level:
          type: string
          title: Trust Level
          default: MEDIUM
        confidentiality:
          type: string
          title: Confidentiality
          default: INTERNAL
        model_profile_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Model Profile Id
        consent_model_egress:
          type: boolean
          title: Consent Model Egress
          default: true
      type: object
      required:
      - files
      title: Body_create_curation_session_api_v1_knowledge_curations_post
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    KnowledgeCurationChatRequest:
      properties:
        instruction:
          type: string
          maxLength: 20000
          minLength: 1
          title: Instruction
        expected_draft_version:
          type: integer
          minimum: 1.0
          title: Expected Draft Version
      type: object
      required:
      - instruction
      - expected_draft_version
      title: KnowledgeCurationChatRequest
    KnowledgeCurationConfirmRequest:
      properties:
        expected_draft_version:
          type: integer
          minimum: 1.0
          title: Expected Draft Version
      type: object
      required:
      - expected_draft_version
      title: KnowledgeCurationConfirmRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
security:
- ApiKey: []